py/examples/ebird-dashboard/requirements.txt
shiny
shinyswatch
pandas
seaborn
ebird.api
